test summation operator
