+++ fenicstools/__init__.py
@@ -6,10 +6,10 @@ try:
 except:
     warnings.warn("Probe/Probes/StatisticsProbe/StatisticsProbes not installed")
 
-# try:
-#     from StructuredGrid import StructuredGrid, ChannelGrid
-# except:
-#     warnings.warn("StructuredGrid/ChannelGrid not installed")
+try:
+    from StructuredGrid import StructuredGrid, ChannelGrid
+except:
+    warnings.warn("StructuredGrid/ChannelGrid not installed")
 
 try:
     from WeightedGradient import weighted_gradient_matrix, compiled_gradient_module
@@ -26,15 +26,15 @@ try:
 except:
     warnings.warn("gauss_divergence/divergence_matrix not installed")
 
-# try:
-#     from Interpolation import interpolate_nonmatching_mesh, interpolate_nonmatching_mesh_any
-# except:
-#     warnings.warn("interpolate_nonmatching_mesh/interpolate_nonmatching_mesh_any not installed")
-# 
-# try:
-#     from DofMapPlotter import DofMapPlotter
-# except:
-#     warnings.warn("DofMapPlotter not installed") # Probably missing dependency
+try:
+    from Interpolation import interpolate_nonmatching_mesh, interpolate_nonmatching_mesh_any
+except:
+    warnings.warn("interpolate_nonmatching_mesh/interpolate_nonmatching_mesh_any not installed")
+
+try:
+    from DofMapPlotter import DofMapPlotter
+except:
+    warnings.warn("DofMapPlotter not installed") # Probably missing dependency
 
 try:
     from ClementInterpolation import clement_interpolate
